python/restore_db/requirements.txt
PyYAML==5.1.2
PyInquirer==1.0.3
PyMySQL==1.0.2